Make prom metrics actually work
